feat: 챌린지 리워드 및 추천시스템 데이터변환 코드(테스트코드에 잇음)

+++ src/main/java/com/hana/api/challenge/dto/request/ChallengeCreateRequest.java
@@ -11,7 +11,7 @@ public class ChallengeCreateRequest {
     private String challengeName;
     private int challengePeriod;
     private String challengeCategory;
-    private Long challengeCost;
+    private int challengeCost;
     private int challengeTargetAmount;
     private List<String> challengers;
 //    private String role;
